refactor: modularize dj feature
