+++ package.json
@@ -21,6 +21,7 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@prisma/client": "5.9.1",
+    "@rollup/rollup-linux-x64-gnu": "^4.20.0",
     "@shadcn/ui": "0.0.4",
     "@supabase/supabase-js": "2.42.0",
     "@upstash/ratelimit": "2.0.6",
@@ -57,6 +58,7 @@
     "ts-node": "10.9.2",
     "tsx": "^4.20.4",
     "typescript": "5.4.5",
+    "vite-tsconfig-paths": "^4.3.2",
     "vitest": "0.34.6"
   },
   "lint-staged": {
