+++ src/app/layout.tsx
@@ -1,12 +1,1 @@
-import '../styles/globals.css';
-import type { ReactNode } from 'react';
-
-export const metadata = { title: 'DJ Requests' };
-
-export default function RootLayout({ children }: { children: ReactNode }) {
-  return (
-    <html lang="en">
-      <body className="min-h-screen bg-gray-900 text-white">{children}</body>
-    </html>
-  );
-}
+export { default, metadata } from '@dj/app/layout';
