+++ src/modules/dj/components/kanban/Card.tsx
