+++ vitest.config.ts
@@ -1,7 +1,9 @@
-import { defineConfig } from "vitest/config";
+import { defineConfig } from 'vitest/config';
+import tsconfigPaths from 'vite-tsconfig-paths';
 
 export default defineConfig({
+  plugins: [tsconfigPaths()],
   test: {
-    environment: "jsdom",
+    environment: 'jsdom',
   },
 });
